correction crash client pub

+++ Serveur.cpp
@@ -12,8 +12,10 @@
 #include <string.h>
 #include <unistd.h>
 #include <setjmp.h>
+#include <errno.h>
 #include "protocole.h" // contient la cle et la structure d'un message
 #include "FichierClient.h" // module de gestion des clients
+#include "Semaphores.h"
 
 int idQ,idShm,idSem;
 int fdPipe[2];
@@ -24,16 +26,15 @@ size_t taille_msg = sizeof(MESSAGE) - sizeof(long);
 void afficheTab();
 
 void handlerSIGINT(int sig);
-
 void handlerSIGCHLD(int sig);
 
 int rechercheTableConnexion(int pid);
-
 int rechercheCaddieTableConnexion(int pid);
 
 int main()
 {
   int index_tab;
+  bool initialisation[6] = {false, false, false, false, false, false};
   // Armement des signaux
 
   struct sigaction B;
@@ -48,23 +49,29 @@ int main()
   sigemptyset(&C.sa_mask);
   sigaction(SIGCHLD, &C, NULL);
 
-  // Creation des ressources
-  // Creation de la file de message
-  fprintf(stderr,"(SERVEUR %d) Creation de la file de messages\n",getpid());
+  // Création des ressources
+  // Création de la file de message
+  fprintf(stderr,"(SERVEUR %d) Création de la file de messages\n",getpid());
   // CLE definie dans protocole.h
   if ((idQ = msgget(CLE,IPC_CREAT | IPC_EXCL | 0600)) == -1){perror("(SERVEUR) Erreur de msgget");exit(EXIT_FAILURE);}
 
   // TO BE CONTINUED
 
-  // Creation du pipe
+  // Création du pipe
+  fprintf(stderr,"(SERVEUR %d) Création du pipe\n",getpid());
   if(pipe(fdPipe)){perror("Erreur de pipe");exit(EXIT_FAILURE);}
   int flags = fcntl(fdPipe[0], F_GETFL);
   flags |= O_NONBLOCK;
   fcntl(fdPipe[0], F_SETFL, flags);
 
+  //Création du sémaphore
+  fprintf(stderr,"(SERVEUR %d) Création du sémaphore\n",getpid());
+  if((idSem = semget(CLE, 1, IPC_CREAT | IPC_EXCL | 0600)) == -1){perror("(SERVEUR) Erreur de semget");exit(EXIT_FAILURE);}
+
+  if(semctl(idSem, 0, SETVAL, 1) == -1){perror("(SERVEUR) Erreur de semctl");exit(EXIT_FAILURE);}
 
   // Initialisation du tableau de connexions
-  tab = (TAB_CONNEXIONS*) malloc(sizeof(TAB_CONNEXIONS)); 
+  tab = (TAB_CONNEXIONS*) malloc(sizeof(TAB_CONNEXIONS));
 
   for (int i=0 ; i<6 ; i++){
     tab->connexions[i].pidFenetre = 0;
@@ -76,10 +83,12 @@ int main()
 
   afficheTab();
 
-  // Creation du processus Publicite (étape 2)
+  // Création du processus Publicite (étape 2)
 
   if((idShm = shmget(CLE, 52, IPC_CREAT | IPC_EXCL | 0600)) == -1){perror("(SERVEUR) Erreur de shmget");exit(EXIT_FAILURE);}
+  
   tab->pidPublicite = fork();
+  
   if(tab->pidPublicite == -1){perror("Erreur de fork");exit(EXIT_FAILURE);}
   else if(tab->pidPublicite == 0){
     char str[10];
@@ -87,7 +96,7 @@ int main()
     if(execl("./Publicite", "Publicite", str, NULL) == -1){perror("Erreur de execl");exit(EXIT_FAILURE);}
   }
 
-  // Creation du processus AccesBD (étape 4)
+  // Création du processus AccesBD (étape 4)
   tab->pidAccesBD = fork();
   if(tab->pidAccesBD == -1){perror("Erreur de fork");exit(EXIT_FAILURE);}
   else if(tab->pidAccesBD == 0){
@@ -105,7 +114,7 @@ int main()
   	fprintf(stderr,"(SERVEUR %d) Attente d'une requete...\n",getpid());
     // Sauvegarde du contexte, définissant où reprendre en cas d'interuption par un signal
     sigsetjmp(contexte, 1);
-    if (msgrcv(idQ, &m, sizeof(MESSAGE)-sizeof(long), 1, 0) == -1){
+    if (msgrcv(idQ, &m, taille_msg, 1, 0) == -1){
       perror("(SERVEUR) Erreur de msgrcv");
       msgctl(idQ, IPC_RMID, NULL);
       exit(EXIT_FAILURE);
@@ -115,6 +124,7 @@ int main()
     {
       case CONNECT :
               fprintf(stderr, "(SERVEUR %d) Requete CONNECT reçue de %d\n", getpid(), m.expediteur);
+              
               for(int i = 0; i<6; i++){
                 if(tab->connexions[i].pidFenetre == 0){
                   tab->connexions[i].pidFenetre = m.expediteur;
@@ -125,11 +135,26 @@ int main()
 
       case DECONNECT :
               fprintf(stderr,"(SERVEUR %d) Requete DECONNECT reçue de %d\n",getpid(),m.expediteur);
+              
               if((index_tab = rechercheTableConnexion(m.expediteur)) != -1)
                 tab->connexions[index_tab].pidFenetre = 0;
               break;
-      case LOGIN :    
+      case LOGIN :
               fprintf(stderr,"(SERVEUR %d) Requete LOGIN reçue de %d : --%d--%s--%s--\n",getpid(),m.expediteur,m.data1,m.data2,m.data3);
+              
+              //Acquisition du sémaphore en no wait, donc si erreur, envoi d'un message busy au client
+              int ret;
+              if((ret = sem_wait(idSem, true)) == -1){
+                reponse.type = m.expediteur;
+                m.requete = 0;
+                reponse.requete = BUSY;
+                reponse.expediteur = getpid();
+                if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
+                kill(m.expediteur, SIGUSR1);
+                break;
+              }
+              fprintf(stderr, "valeur retour wait : %d\n", ret);
+
               reponse.data1 = 1;
 
               // Nouveau Client
@@ -191,10 +216,22 @@ int main()
 
               if (msgsnd(idQ, &reponse, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
               kill(m.expediteur, SIGUSR1);
+              
+              // Libération du semaphore
+              sem_signal(idSem);
               break;
 
       case LOGOUT :
               fprintf(stderr,"(SERVEUR %d) Requete LOGOUT reçue de %d\n",getpid(), m.expediteur);
+
+              //Acquisition du sémaphore en no wait, donc si erreur, envoi d'un message busy au client
+              if(sem_wait(idSem, true) == -1){
+                reponse.type = m.expediteur;
+                reponse.requete = BUSY;
+                reponse.expediteur = getpid();
+                if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
+                break;
+              }
 
               if((index_tab = rechercheTableConnexion(m.expediteur))!= -1){
                 strcpy(tab->connexions[index_tab].nom, "");
@@ -205,12 +242,21 @@ int main()
                 }
                 tab->connexions[index_tab].pidCaddie = 0;
               }
+
+              // Libération du semaphore
+              sem_signal(idSem);
               break;
 
       case UPDATE_PUB : 
+              fprintf(stderr,"(SERVEUR %d) Requete UPDATE_PUB reçue de %d\n",getpid(), m.expediteur);
+              
               for(int i = 0; i<6; i++){
-                if(tab->connexions[i].pidFenetre != 0)
-                  kill(tab->connexions[i].pidFenetre, SIGUSR2);
+                if(tab->connexions[i].pidFenetre != 0){
+                  if(initialisation[i])
+                    kill(tab->connexions[i].pidFenetre, SIGUSR2);
+                  else
+                    initialisation[i] = true;
+                }
               }
 
               break;
@@ -218,6 +264,15 @@ int main()
       case CONSULT :  
               fprintf(stderr,"(SERVEUR %d) Requete CONSULT reçue de %d\n",getpid(),m.expediteur);
               
+              //Acquisition du sémaphore en no wait, donc si erreur, envoi d'un message busy au client
+              if(sem_wait(idSem, true) == -1){
+                reponse.type = m.expediteur;
+                reponse.requete = BUSY;
+                reponse.expediteur = getpid();
+                if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
+                break;
+              }              
+
               index_tab = rechercheTableConnexion(m.expediteur);
           
               if(index_tab != -1){
@@ -225,52 +280,109 @@ int main()
 
                 if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
               }
+
+              // Libération du semaphore
+              sem_signal(idSem);
               break;
 
       case ACHAT :    
               fprintf(stderr,"(SERVEUR %d) Requete ACHAT reçue de %d\n",getpid(),m.expediteur);
+              
+              //Acquisition du sémaphore en no wait, donc si erreur, envoi d'un message busy au client
+              if(sem_wait(idSem, true) == -1){
+                reponse.type = m.expediteur;
+                reponse.requete = BUSY;
+                reponse.expediteur = getpid();
+                if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
+                break;
+              }
+
+              index_tab = rechercheTableConnexion(m.expediteur);
+              if(index_tab != -1){
+                m.type = tab->connexions[index_tab].pidCaddie;
+                if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
+              }
+
+              // Libération du semaphore
+              sem_signal(idSem);
+              break;
+
+      case CADDIE :
+              fprintf(stderr,"(SERVEUR %d) Requete CADDIE reçue de %d\n",getpid(),m.expediteur);
+              
+              //Acquisition du sémaphore en no wait, donc si erreur, envoi d'un message busy au client
+              if(sem_wait(idSem, true) == -1){
+                reponse.type = m.expediteur;
+                reponse.requete = BUSY;
+                reponse.expediteur = getpid();
+                if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
+                break;
+              }
               
               index_tab = rechercheTableConnexion(m.expediteur);
               if(index_tab != -1){
                 m.type = tab->connexions[index_tab].pidCaddie;
                 if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
               }
+
+              // Libération du semaphore
+              sem_signal(idSem);
               break;
 
-      case CADDIE :
-              fprintf(stderr,"(SERVEUR %d) Requete CADDIE reçue de %d\n",getpid(),m.expediteur);
+      case CANCEL :
+              fprintf(stderr,"(SERVEUR %d) Requete CANCEL reçue de %d\n",getpid(),m.expediteur);
+              
+              //Acquisition du sémaphore en no wait, donc si erreur, envoi d'un message busy au client
+              if(sem_wait(idSem, true) == -1){
+                reponse.type = m.expediteur;
+                reponse.requete = BUSY;
+                reponse.expediteur = getpid();
+                if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
+                break;
+              }
+
               index_tab = rechercheTableConnexion(m.expediteur);
               if(index_tab != -1){
                 m.type = tab->connexions[index_tab].pidCaddie;
                 if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
               }
 
-              break;
-
-      case CANCEL :
-              fprintf(stderr,"(SERVEUR %d) Requete CANCEL reçue de %d\n",getpid(),m.expediteur);
-              index_tab = rechercheTableConnexion(m.expediteur);
-              if(index_tab != -1){
-                m.type = tab->connexions[index_tab].pidCaddie;
-                if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
-              }       
+              // Libération du semaphore
+              sem_signal(idSem);    
               break;
 
       case CANCEL_ALL :
               fprintf(stderr,"(SERVEUR %d) Requete CANCEL_ALL reçue de %d\n",getpid(),m.expediteur);
+              
+              //Acquisition du sémaphore en no wait, donc si erreur, envoi d'un message busy au client
+              if(sem_wait(idSem, true) == -1){
+                reponse.type = m.expediteur;
+                reponse.requete = BUSY;
+                reponse.expediteur = getpid();
+                if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
+                break;
+              }
+
               index_tab = rechercheTableConnexion(m.expediteur);
               if(index_tab != -1){
                 m.type = tab->connexions[index_tab].pidCaddie;
                 if (msgsnd(idQ, &m, taille_msg, 0) == -1){perror("Erreur de msgsnd");exit(EXIT_FAILURE);}
-              }       
+              }
+
+              // Libération du semaphore
+              sem_signal(idSem);    
               break;
 
       case PAYER : // TO DO
               fprintf(stderr,"(SERVEUR %d) Requete PAYER reçue de %d\n",getpid(),m.expediteur);
+              
+              // Libération du semaphore
+              // sem_signal(idSem);
               break;
 
       case NEW_PUB :  // TO DO
               fprintf(stderr,"(SERVEUR %d) Requete NEW_PUB reçue de %d\n",getpid(),m.expediteur);
+              
               break;
     }
     afficheTab();
@@ -292,16 +404,28 @@ void afficheTab()
 ////////////////////////////////////////////////////////////////////////////////////
 
 void handlerSIGINT(int sig){
+  errno = 0;
   // Supprime la file de messages
-  if (msgctl(idQ, IPC_RMID, NULL) == -1){perror("Erreur de msgctl");}
+  if (msgctl(idQ, IPC_RMID, NULL) == -1)
+    perror("(SERVEUR) Erreur de msgctl");
 
   // Supprime la mémoire partagée
-  if (shmctl(idShm, IPC_RMID, NULL) == -1){perror("Erreur de shmctl");}
+  if (shmctl(idShm, IPC_RMID, NULL) == -1)
+    perror("(SERVEUR) Erreur de shmctl");
 
   // Ferme les descripteurs du pipe sur le serveur
-  close(fdPipe[0]);
-  close(fdPipe[1]);
+  if(close(fdPipe[0]) == -1)
+    perror("(SERVEUR) Erreur de close");
+  if(close(fdPipe[1]) == -1)
+    perror("(SERVEUR) Erreur de close");
 
+  // Supprime le sémaphore
+  if (semctl(idSem, 0, IPC_RMID) == -1)
+    perror("(SERVEUR) Erreur de semctl");
+
+  if(errno)
+    exit(EXIT_FAILURE);
+  
   exit(EXIT_SUCCESS);
 }
